requirements.txt
accelerate==1.10.1
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aioitertools==0.13.0
aiosignal==1.3.2
alembic==1.18.4
anaconda-anon-usage==0.7.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.12.1
archspec==0.2.3
asttokens==3.0.0
attrs==25.3.0
av==16.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
boltons==24.1.0
boto3==1.40.61
botocore==1.40.61
Brotli==1.0.9
cachetools==7.0.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
cloudpickle==3.1.2
comm==0.2.2
conda==25.3.1
conda-anaconda-telemetry==0.1.2
conda-anaconda-tos==0.1.3
conda-content-trust==0.2.0
conda-libmamba-solver==25.4.0
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.3
cryptography==44.0.1
cycler==0.12.1
databricks-sdk==0.100.0
datasets==4.8.2
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
executing==2.2.0
fastapi==0.135.1
fastuuid==0.14.0
filelock==3.18.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.0
frozendict==2.4.2
frozenlist==1.6.0
fsspec==2025.3.0
gepa==0.1.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.7.1
idna==3.7
ijson==3.5.0
importlib_metadata==8.7.1
inspect_ai==0.3.200.dev5+g67da5fe22
inspect_evals==0.5.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==2.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.5.0
libmambapy==2.0.5
linkify-it-py==2.1.0
litellm==1.82.4
Mako==1.3.10
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.0
menuinst==2.2.0
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mmh3==5.2.1
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
nest-asyncio==1.6.0
nest-asyncio2==1.7.2
networkx==3.5
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==2.29.0
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib_abc==0.5.2
pexpect==4.9.0
pillow==11.3.0
pip==25.0
platformdirs==4.3.7
playwright==1.58.0
pluggy==1.5.0
prettytable==3.17.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.33.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycosat==0.6.6
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.3.0
s3transfer==0.14.0
safetensors==0.5.3
scikit-learn==1.8.0
scipy==1.17.1
semver==3.0.4
sentry-sdk==2.55.0
setuptools==78.1.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
textual==8.1.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.0
toml==0.10.2
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0.dev0
triton==3.3.0
truststore==0.10.0
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uc-micro-py==2.0.0
universal_pathlib==0.3.10
urllib3==2.3.0
uv==0.10.11
uvicorn==0.42.0
wandb==0.25.1
wcwidth==0.2.13
Werkzeug==3.1.6
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.0
zipfile-zstd==0.0.4
zipp==3.23.0
zstandard==0.23.0
